news page implemented

+++ app/Controllers/AdminAvisController.php
@@ -16,48 +16,52 @@ class AdminAvisController extends Controller{
 
     public function loadPage($data)
     {
-        //models declaration area
-        $avisMarques = new AvisMarqueModel();
-        $avisVehicules = new AvisVehiculeModel();
+        if(isset($_SESSION['admin-authenticated'])){
+            //models declaration area
+            $avisMarques = new AvisMarqueModel();
+            $avisVehicules = new AvisVehiculeModel();
 
-        if(isset($data[1])){
-            switch ($data[1]){
-                //MODELS
-                case 'ValiderMarque' :
-                    $avis = $avisMarques->get($data[2]);
-                    $avis['valide'] = 1;
-                    $avisMarques->update($avis);
-                    break;
-                case 'RefuserMarque' :
-                    $avisMarques->delete($data[2]);
-                    break;
-                case 'ValiderVehicule' :
-                    $avis = $avisVehicules->get($data[2]);
-                    $avis['valide'] = 1;
-                    $avisVehicules->update($avis);
-                    break;
-                case 'RefuserVehicule' :
-                    $avisVehicules->delete($data[2]);
-                    break;
+            if(isset($data[1])){
+                switch ($data[1]){
+                    //MODELS
+                    case 'ValiderMarque' :
+                        $avis = $avisMarques->get($data[2]);
+                        $avis['valide'] = 1;
+                        $avisMarques->update($avis);
+                        break;
+                    case 'RefuserMarque' :
+                        $avisMarques->delete($data[2]);
+                        break;
+                    case 'ValiderVehicule' :
+                        $avis = $avisVehicules->get($data[2]);
+                        $avis['valide'] = 1;
+                        $avisVehicules->update($avis);
+                        break;
+                    case 'RefuserVehicule' :
+                        $avisVehicules->delete($data[2]);
+                        break;
 
+                }
             }
+
+
+            //views declaration area
+            $head = new HeadView();
+            $avisMarqueTable = new AdminMarqueAvisTableView();
+            $avisVehiculeTable = new AdminVehiculeAvisTableView();
+            $bottom = new BottomView();
+
+
+            $avisMarquesData = $avisMarques->getAll();
+            $avisVehiculeData = $avisVehicules->getAll();
+
+            $head->show(null);
+            $avisMarqueTable->show($avisMarquesData);
+            $avisVehiculeTable->show($avisVehiculeData);
+            $bottom->show(null);
+        }else{
+            header("Location: http://localhost/cars_website_scratch_version/admin/AdminAuth");
         }
-
-
-        //views declaration area
-        $head = new HeadView();
-        $avisMarqueTable = new AdminMarqueAvisTableView();
-        $avisVehiculeTable = new AdminVehiculeAvisTableView();
-        $bottom = new BottomView();
-
-
-        $avisMarquesData = $avisMarques->getAll();
-        $avisVehiculeData = $avisVehicules->getAll();
-
-        $head->show(null);
-        $avisMarqueTable->show($avisMarquesData);
-        $avisVehiculeTable->show($avisVehiculeData);
-        $bottom->show(null);
 
     }
 }
