+++ app/Controllers/AdminMarqueController.php
@@ -16,49 +16,52 @@ class AdminMarqueController extends Controller{
 
     public function loadPage($data)
     {
-        //models declaration area
-        $marques = new MarqueModel();
+        if(isset($_SESSION['admin-authenticated'])) {
+            //models declaration area
+            $marques = new MarqueModel();
 
-        //views declaration area
-        $head = new HeadView();
-        $form = new AdminMarqueFormView();
-        $bottom = new BottomView();
+            //views declaration area
+            $head = new HeadView();
+            $form = new AdminMarqueFormView();
+            $bottom = new BottomView();
 
-        switch ($data[1]){
-            //MODELS
-            case 'Modifier' :
-                //binding area
-                $marqueData = $marques->get($data[2]);
+            switch ($data[1]) {
+                //MODELS
+                case 'Modifier' :
+                    //binding area
+                    $marqueData = $marques->get($data[2]);
 
-                if(array_key_exists("MarqueButton",$_POST))
-                {
-                    $marqueData = $marques->get($_POST['marque_id']);
-                    $marques->update($_POST);
+                    if (array_key_exists("MarqueButton", $_POST)) {
+                        $marqueData = $marques->get($_POST['marque_id']);
+                        $marques->update($_POST);
+                        header("Location: http://localhost/cars_website_scratch_version/admin/AdminMV");
+                    }
+
+                    //display area
+                    $head->show(null);
+                    $form->show($marqueData);
+                    $bottom->show(null);
+                    break;
+                case 'Supprimer' :
+                    $marques->delete($data[2]);
                     header("Location: http://localhost/cars_website_scratch_version/admin/AdminMV");
-                }
+                    break;
+                case 'Ajouter' :
+                    //inserting area
+                    if (array_key_exists("MarqueButton", $_POST)) {
+                        $marques->insert($_POST);
+                        header("Location: http://localhost/cars_website_scratch_version/admin/AdminMV");
+                    }
 
-                //display area
-                $head->show(null);
-                $form->show($marqueData);
-                $bottom->show(null);
-                break;
-            case 'Supprimer' :
-                $marques->delete($data[2]);
-                header("Location: http://localhost/cars_website_scratch_version/admin/AdminMV");
-                break;
-            case 'Ajouter' :
-                //inserting area
-                if(array_key_exists("MarqueButton",$_POST))
-                {
-                    $marques->insert($_POST);
-                    header("Location: http://localhost/cars_website_scratch_version/admin/AdminMV");
-                }
-
-                //display area
-                $head->show(null);
-                $form->show(null);
-                $bottom->show(null);
-                break;
+                    //display area
+                    $head->show(null);
+                    $form->show(null);
+                    $bottom->show(null);
+                    break;
+            }
+        }else{
+            header("Location: http://localhost/cars_website_scratch_version/admin/AdminAuth");
         }
+
     }
 }
